libalsa-intf: Initialize the data size of header to 0 for compressed.
Issue with aplay is that it was not reading the length in header.
when silence files are created using third party pcm generators, it
has some meta data at the end of file after all zeros. So aplay
will send that data also. This behavior is modified now to avoid
pops in the end of the file by keeping track of the header data size.
As for the compressed no such track is needed so initializing with
the data size of header to 0

Change-Id: Id880a92fb9fc8e8c7c74c900baf2911546116073